application.jsをassetsへ移動
